fix(renderer): keep interaction surfaces out of startup chunk
