+++ code-review/architecture.md
@@ -125,7 +125,7 @@ Opening a folder is a navigation action first and a preparation action second. O
 
 Each opened folder can carry a short optional description in app config. The description is orientation metadata for humans and Agents: it explains what the folder is for, but it is not indexed content and it does not define access scope. It can be written by the user first and later generated or refreshed by AI. Removing a folder from "Your Folders" removes its description with the folder membership record.
 
-Renderer state orchestration lives under `web-src/src/store/`. `state.ts` is the stable state-model and action-contract facade, `stateHelpers.ts` owns reusable pure transitions and layout bounds, and `stateReducer.ts` applies the action union. `useActiveFolderWorkspace.ts` is the active-folder workspace module: it owns folder, document, retrieval, and lifecycle ordering, including polling, focus reconciliation, binary refresh, and context-release persistence. It composes `useDocumentActions.ts`, `useFileActions.ts`, `useSearchActions.ts`, and `useFolderActions.ts` as private implementation details; presentation callers must use the workspace interface rather than coordinate those actions or freshness guards themselves. The interface must retain its identity when its commands are unchanged: document surfaces use it in effects, and a replacement while Find is open re-registers the controller and can create a render loop. Its lifecycle seam has two adapters: Electron context-release callbacks and the browser unload fallback. `AppContext.tsx` composes that workspace with shell-owned chat, feedback, and Find interaction. `useFindActions.ts` and `useFeedbackActions.ts` remain shell presentation protocols; search-hit navigation remains in the workspace. `indexStatusRequest.ts` is the narrow request-lifecycle seam that classifies status responses against folder-transition state without owning scheduling, recovery, or renderer state. `web-src/src/components/MainPane.tsx` dynamically imports format- and mode-specific heavy viewers/editors, including audio, PDF, DOCX, Markdown preview, and the Markdown editor, so the initial renderer chunk carries the common browsing surface first. `web-src/src/App.tsx` also dynamically imports the chat pane; the sidebar's New Chat entry stays in the initial shell, while Agent transcript rendering and the CodeMirror mention composer load only after chat is opened. `scripts/check-renderer-chunks.mjs` requires the five designated dynamic entries and caps the entry chunk plus its recursive static imports at 400 KiB. `web-src/src/components/ErrorBoundary.tsx` retries each dynamic import once and contains a persistent failure inside the affected chat or document surface; a changed document identity/version or chat surface clears that local failure, while the root boundary remains the final recovery path for unrelated renderer errors. `web-src/src/components/ChatPane.tsx` keeps tab navigation outside one boundary per mounted Agent session, so a render failure in one tab cannot hide the controls needed to switch or close it.
+Renderer state orchestration lives under `web-src/src/store/`. `state.ts` is the stable state-model and action-contract facade, `stateHelpers.ts` owns reusable pure transitions and layout bounds, and `stateReducer.ts` applies the action union. `useActiveFolderWorkspace.ts` is the active-folder workspace module: it owns folder, document, retrieval, and lifecycle ordering, including polling, focus reconciliation, binary refresh, and context-release persistence. It composes `useDocumentActions.ts`, `useFileActions.ts`, `useSearchActions.ts`, and `useFolderActions.ts` as private implementation details; presentation callers must use the workspace interface rather than coordinate those actions or freshness guards themselves. The interface must retain its identity when its commands are unchanged: document surfaces use it in effects, and a replacement while Find is open re-registers the controller and can create a render loop. Its lifecycle seam has two adapters: Electron context-release callbacks and the browser unload fallback. `AppContext.tsx` composes that workspace with shell-owned chat, feedback, and Find interaction. `useFindActions.ts` and `useFeedbackActions.ts` remain shell presentation protocols; search-hit navigation remains in the workspace. `indexStatusRequest.ts` is the narrow request-lifecycle seam that classifies status responses against folder-transition state without owning scheduling, recovery, or renderer state. `web-src/src/components/MainPane.tsx` dynamically imports format- and mode-specific heavy viewers/editors, including audio, PDF, DOCX, Markdown preview, and the Markdown editor, so the initial renderer chunk carries the common browsing surface first. `web-src/src/App.tsx` also dynamically imports the chat pane; the sidebar's New Chat entry stays in the initial shell, while Agent transcript rendering and the CodeMirror mention composer load only after chat is opened. The sidebar dynamically imports the search panel when its activity-bar view opens and the document-outline list when a Markdown outline is visible. Its eager files surface checks semantic-indexing and unsupported-file state before importing those normally absent disclosures. Quick Open keeps its shortcut listener in the initial shell but dynamically imports ranking and picker rendering after an open request, so the first shortcut cannot race module loading. Context-menu and image-lightbox implementations also load only after their state-backed requests exist. The embedder-key gate keeps folder probing and overlay ownership eager, then loads its form only when setup must open. `scripts/check-renderer-chunks.mjs` requires these twelve designated dynamic entries and caps the entry chunk plus its recursive static imports at 400 KiB. `web-src/src/components/ErrorBoundary.tsx` retries each dynamic import once and contains a persistent failure inside the affected interaction, chat, search, Quick Open, or document surface; a changed request, document identity/version, chat surface, or sidebar view clears that local failure, while the root boundary remains the final recovery path for unrelated renderer errors. `web-src/src/components/ChatPane.tsx` keeps tab navigation outside one boundary per mounted Agent session, so a render failure in one tab cannot hide the controls needed to switch or close it.
 
 Toast lifecycle is intentionally not reducer state: the shared Base UI toast
 manager owns timeout, close, announcement priority, and viewport navigation.
